+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [`main`](https://github.com/walterra/d3-milestones/tree/main)
 
+- Added support for ordinal scales as an alternative to time scales. ([#15](https://github.com/walterra/d3-milestones/issues/15))
 - Add `renderCallback()` method to apply customizations after rendering is complete. ([#79](https://github.com/walterra/d3-milestones/issues/79))
 - Added support for custom HTML ID attributes for milestone elements. ([#78](https://github.com/walterra/d3-milestones/issues/78))
 - Added WebP format to the list of supported image formats.